change-ai-bot-main/requirements.txt
aerich==0.7.1 ; python_version >= "3.8" and python_version < "4.0"
aiofiles==23.1.0 ; python_version >= "3.8" and python_version < "4.0"
aiogram-fastapi-server==0.1.1 ; python_version >= "3.8" and python_version < "4.0"
aiogram==3.0.0b7 ; python_version >= "3.8" and python_version < "4.0"
aiohttp==3.8.4 ; python_version >= "3.8" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.8" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.8" and python_version < "4.0"
asyncpg==0.27.0 ; python_version >= "3.8" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.8" and python_version < "4.0"
babel==2.11.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
dictdiffer==0.9.0 ; python_version >= "3.8" and python_version < "4.0"
fastapi==0.92.0 ; python_version >= "3.8" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.8" and python_version < "4.0"
magic-filter==1.0.9 ; python_version >= "3.8" and python_version < "4.0"
marshmallow-dataclass==8.5.11 ; python_version >= "3.8" and python_version < "4.0"
marshmallow==3.19.0 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.8" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.8" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.8" and python_version < "4.0"
pydantic==1.10.5 ; python_version >= "3.8" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.8" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.8" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.8" and python_version < "4.0"
redis==4.5.1 ; python_version >= "3.8" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
starlette==0.25.0 ; python_version >= "3.8" and python_version < "4.0"
tomlkit==0.11.6 ; python_version >= "3.8" and python_version < "4.0"
tortoise-orm==0.19.3 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.8" and python_version < "4.0"
typing-inspect==0.8.0 ; python_version >= "3.8" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.8" and python_version < "4.0"
